refactor(StructureAuthoring): remove unused imports and restructure routing
