+++ src/assets/wise5/authoringTool/structure/structure-authoring.module.ts
@@ -1,22 +1,23 @@
 import { NgModule } from '@angular/core';
-import { StudentTeacherCommonModule } from '../../../../app/student-teacher-common.module';
 import { JigsawComponent } from './jigsaw/jigsaw.component';
 import { SelfDirectedInvestigationComponent } from './self-directed-investigation/self-directed-investigation.component';
 import { KiCycleUsingOerComponent } from './ki-cycle-using-oer/ki-cycle-using-oer.component';
 import { PeerReviewAndRevisionComponent } from './peer-review-and-revision/peer-review-and-revision.component';
-import { RouterModule } from '@angular/router';
-import { StructureAuthoringRoutingModule } from './structure-authoring-routing.module';
+import { RouterModule, Routes } from '@angular/router';
+
+const routes: Routes = [
+  {
+    path: '',
+    children: [
+      { path: 'jigsaw', component: JigsawComponent },
+      { path: 'ki-cycle-using-oer', component: KiCycleUsingOerComponent },
+      { path: 'peer-review-and-revision', component: PeerReviewAndRevisionComponent },
+      { path: 'self-directed-investigation', component: SelfDirectedInvestigationComponent }
+    ]
+  }
+];
 
 @NgModule({
-    imports: [RouterModule, StudentTeacherCommonModule, StructureAuthoringRoutingModule, JigsawComponent,
-        KiCycleUsingOerComponent,
-        PeerReviewAndRevisionComponent,
-        SelfDirectedInvestigationComponent],
-    exports: [
-        JigsawComponent,
-        KiCycleUsingOerComponent,
-        PeerReviewAndRevisionComponent,
-        SelfDirectedInvestigationComponent
-    ]
+  imports: [RouterModule.forChild(routes)]
 })
 export class StructureAuthoringModule {}
